Extract the factory services in own service yml
