+++ src/DependencyInjection/MetaModelsAttributeTranslatedFileExtension.php
@@ -36,6 +36,7 @@ class MetaModelsAttributeTranslatedFileExtension extends Extension
     public function load(array $configs, ContainerBuilder $container)
     {
         $loader = new Loader\YamlFileLoader($container, new FileLocator(__DIR__ . '/../Resources/config'));
+        $loader->load('factory.yml');
         $loader->load('services.yml');
     }
 }
